# docs/api/hermes_openapi.yaml
openapi: 3.0.3
info:
  title: Hermes
  description: |-
    Hermes is a system of recommendations and tourist planner, to facilitate the most tedious task of any trip: **planning it**.

    This page documents Hermes REST API that is used by the frontend.

    Most of the endpoints need authentication. This is done by a cookie called `hermesSession`.
  version: 1.0.3
  contact:
    name: Álvaro Galisteo
    url: 'https://alvaro.galisteo.me'
    email: hermes@galisteo.me
  license:
    name: GPL 3.0
    url: 'https://github.com/SrGMC/hermes/blob/main/LICENSE'
  termsOfService: 'https://travelhermes.com/terms/'
servers:
  - url: 'https://travelhermes.com'
components:
  securitySchemes:
    hermesSession:
      description: 'When authentication is needed, the `hermesSession` session cookie is required. This has a value of a 255 char string.'
      type: apiKey
      in: cookie
      name: hermesSession
...
paths:
...
  /api/ratings/create:
    post:
      security:
        - hermesSession: []
      tags:
        - Ratings
      summary: Create a new rating
      description: |-
        Creates a new rating.

        Rating scale goes from 1 to 5.
      operationId: createANewRating
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                placeId:
                  type: number
                  example: 25
                rating:
                  type: number
                  example: 3
                  minimum: 1
                  maximum: 5
              required:
                - placeId
                - rating
            example:
              placeId: 25
              rating: 3
      responses:
        '200':
          description: Successful creation
          content: {}
        '400':
          description: Missing or bad attributes
          content: {}
        '404':
          description: Place not found
          content: {}
  /api/ratings/delete:
    post:
      security:
        - hermesSession: []
      tags:
        - Ratings
      summary: Delete a rating
      description: 'Deletes an existing rating. Ratings are identified by user ID (implicit, from cookie) and place ID (required).'
      operationId: deleteARating
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                placeId:
                  type: number
                  example: 45
              required:
                - placeId
            example:
              placeId: 45
      responses:
        '200':
          description: Successful deletion
          content: {}
        '400':
          description: Missing or bad attribute
          content: {}
        '404':
          description: Rating or place not found
          content: {}
...
  '/api/ratings/search/{query}':
    get:
      security:
        - hermesSession: []
      tags:
        - Ratings
      summary: Search for places not rated by user
      description: Returns a list of places in which a rating by the user does not exist.
      operationId: searchForPlacesNotRatedByUser
      responses:
        '200':
          description: Search multiple places
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    example:
                      - description: 'Importante museo de colecciones de arte español del siglo XX con obras de Gris, Dalí, Picasso y Miró.'
                        id: 41
                        images: 1
                        name: Museo Nacional Centro de Arte Reina Sofía (MNCARS)
                        placeUrl: 'http://www.museoreinasofia.es'
                        rating: 1
                        wikipedia: 'es:Museo Nacional Centro de Arte Reina Sofía'
                      - description: 'Colección de obras de arte de los siglos XII a XIX y obras maestras de Velázquez, Goya y el Greco.'
                        id: 43
                        images: 1
                        name: Museo Nacional del Prado
                        placeUrl: 'http://www.museodelprado.es'
                        rating: 5
                        wikipedia: 'es:Museo del Prado'
                      - description: 'Museo dedicado a la divulgación científica y tecnológica, y a la conservación. Posee una colección de más de 15000 instrumentos científicos que datan desde el siglo XVI hasta la actualidad.'
                        id: 49
                        images: 1
                        name: Museo Nacional de Ciencia y Tecnología
                        placeUrl: 'http://www.muncyt.es/'
                        rating: 5
                        wikipedia: 'es:Museo_Nacional_de_Ciencia_y_Tecnología'
                      - description: 'Museo que alberga obras maestras europeas del s. XIII al XX, desde el Renacimiento hasta el Arte pop.'
                        id: 52
                        images: 1
                        name: Museo Thyssen-Bornemisza
                        placeUrl: 'http://www.museothyssen.org'
                        rating: 4
                        wikipedia: 'es:Museo Thyssen-Bornemisza'
                      - description: 'Museo de cera con figuras de famosos, un simulador espacial, un tren fantasma y una proyección sobre Historia.'
                        id: 169
                        images: 1
                        name: Museo de Cera
                        placeUrl: 'http://www.museoceramadrid.com'
                        rating: 1
                        wikipedia: 'es:Museo_de_Cera_de_Madrid'
                      - description: 'Museo en un edificio neoclásico con obras antiguas y modernas, como arte rupestre y cerámica islámica.'
                        id: 170
                        images: 2
                        name: Museo Arqueológico Nacional (MAN)
                        placeUrl: 'http://man.mcu.es'
                        rating: 5
                        wikipedia: 'es:Museo Arqueológico Nacional (España)'
                    items:
                      type: object
                      properties:
                        description:
                          type: string
                          example: Galería de esculturas y arte contemporáneo
                        id:
                          type: number
                          example: 25
                        images:
                          type: number
                          example: 0
                          minimum: 0
                          description: |-
                            Number of images that the place has.
                            Images are stored in `/assets/places/{placeId}/{imageCount}.jpg`
                        name:
                          type: string
                          minLength: 1
                          example: Galería Ponce+Robles
                        placeUrl:
                          type: string
                          example: 'https://poncerobles.com/es/'
                          format: uri
                        rating:
                          type: number
                          example: 3
                          minimum: 1
                          maximum: 5
                        wikipedia:
                          type: string
                          nullable: true
                          example: null
                          pattern: 'es:.+'
                          minLength: 1
              examples:
                Search multiple places:
                  value:
                    result:
                      - description: 'Importante museo de colecciones de arte español del siglo XX con obras de Gris, Dalí, Picasso y Miró.'
                        id: 41
                        images: 1
                        name: Museo Nacional Centro de Arte Reina Sofía (MNCARS)
                        placeUrl: 'http://www.museoreinasofia.es'
                        rating: 1
                        wikipedia: 'es:Museo Nacional Centro de Arte Reina Sofía'
                      - description: 'Colección de obras de arte de los siglos XII a XIX y obras maestras de Velázquez, Goya y el Greco.'
                        id: 43
                        images: 1
                        name: Museo Nacional del Prado
                        placeUrl: 'http://www.museodelprado.es'
                        rating: 5
                        wikipedia: 'es:Museo del Prado'
                      - description: 'Museo dedicado a la divulgación científica y tecnológica, y a la conservación. Posee una colección de más de 15000 instrumentos científicos que datan desde el siglo XVI hasta la actualidad.'
                        id: 49
                        images: 1
                        name: Museo Nacional de Ciencia y Tecnología
                        placeUrl: 'http://www.muncyt.es/'
                        rating: 5
                        wikipedia: 'es:Museo_Nacional_de_Ciencia_y_Tecnología'
                      - description: 'Museo que alberga obras maestras europeas del s. XIII al XX, desde el Renacimiento hasta el Arte pop.'
                        id: 52
                        images: 1
                        name: Museo Thyssen-Bornemisza
                        placeUrl: 'http://www.museothyssen.org'
                        rating: 4
                        wikipedia: 'es:Museo Thyssen-Bornemisza'
                      - description: 'Museo de cera con figuras de famosos, un simulador espacial, un tren fantasma y una proyección sobre Historia.'
                        id: 169
                        images: 1
                        name: Museo de Cera
                        placeUrl: 'http://www.museoceramadrid.com'
                        rating: 1
                        wikipedia: 'es:Museo_de_Cera_de_Madrid'
                      - description: 'Museo en un edificio neoclásico con obras antiguas y modernas, como arte rupestre y cerámica islámica.'
                        id: 170
                        images: 2
                        name: Museo Arqueológico Nacional (MAN)
                        placeUrl: 'http://man.mcu.es'
                        rating: 5
                        wikipedia: 'es:Museo Arqueológico Nacional (España)'
    parameters:
      - name: query
        in: path
        required: true
        schema:
          type: string
          example: museo de
...
security:
  - hermesSession: []